@@ -27,15 +27,22 @@
     },
     "require": {
         "php": "^8.1",
-        "aeviiq/collection": "^3.6",
+        "aeviiq/collection": "^4.0",
         "aeviiq/storage-manager": "^6.0",
-        "aeviiq/enum": "^3.0",
+        "aeviiq/enum": "^4.0",
         "symfony/form": "^6.0",
         "symfony/http-foundation": "^6.0",
-        "symfony/event-dispatcher-contracts": "^2.0"
+        "symfony/event-dispatcher-contracts": "^2.0|^3.0"
     },
     "require-dev": {
-        "phpunit/phpunit": "^9.5"
+        "phpunit/phpunit": "^10.0",
+        "vimeo/psalm": "^5.4",
+        "phpstan/phpstan": "^1.10",
+        "psalm/plugin-phpunit": "^0.18.4",
+        "thecodingmachine/phpstan-strict-rules": "^1.0",
+        "phpstan/phpstan-strict-rules": "^1.5",
+        "phpstan/phpstan-phpunit": "^1.3",
+        "symfony/var-dumper": "^6.4"
     },
     "scripts": {
         "phpunit": [
