@@ -1,4 +1,6 @@
-<?php declare(strict_types=1);
+<?php
+
+declare(strict_types=1);
 
 namespace Aeviiq\FormFlow\Exception;
 
@@ -6,15 +8,9 @@ use Aeviiq\FormFlow\FormFlowInterface;
 
 final class TransitionException extends \RuntimeException implements ExceptionInterface
 {
-    /**
-     * @var FormFlowInterface
-     */
-    private $flow;
-
-    public function __construct(FormFlowInterface $flow, string $message, int $code = 0, \Throwable $previous = null)
+    public function __construct(private readonly FormFlowInterface $flow, string $message, int $code = 0, \Throwable $previous = null)
     {
         parent::__construct($message, $code, $previous);
-        $this->flow = $flow;
     }
 
     /**
