@@ -16,8 +16,8 @@ final class DefinitionTest extends TestCase
     public function testDefinition(): void
     {
         $definintion = new Definition('some-name', stdClass::class, new StepCollection([
-            $this->createStub(StepInterface::class),
-            $this->createStub(StepInterface::class),
+            self::createStub(StepInterface::class),
+            self::createStub(StepInterface::class),
         ]), ['group-1', 'group-2']);
 
         self::assertSame('some-name', $definintion->getName());
@@ -29,32 +29,32 @@ final class DefinitionTest extends TestCase
     public function testConstructWithInvalidName(): void
     {
         $this->expectException(InvalidArgumentException::class);
-        $this->expectDeprecationMessage('The definition name cannot be empty');
+        $this->expectExceptionMessage('The definition name cannot be empty');
 
         new Definition('', stdClass::class, new StepCollection([
-            $this->createStub(StepInterface::class),
-            $this->createStub(StepInterface::class),
+            self::createStub(StepInterface::class),
+            self::createStub(StepInterface::class),
         ]), ['group-1', 'group-2']);
     }
 
     public function testConstructWithInvalidExpectedInstance(): void
     {
         $this->expectException(InvalidArgumentException::class);
-        $this->expectDeprecationMessage('The expected instance must be an existing class or interface.');
+        $this->expectExceptionMessage('The expected instance must be an existing class or interface.');
 
         new Definition('some-name', 'some-non-existent-class', new StepCollection([
-            $this->createStub(StepInterface::class),
-            $this->createStub(StepInterface::class),
+            self::createStub(StepInterface::class),
+            self::createStub(StepInterface::class),
         ]), ['group-1', 'group-2']);
     }
 
     public function testConstructWithInvalidStepCount(): void
     {
         $this->expectException(InvalidArgumentException::class);
-        $this->expectDeprecationMessage('A flow must consist of at least 2 steps.');
+        $this->expectExceptionMessage('A flow must consist of at least 2 steps.');
 
         new Definition('some-name', stdClass::class, new StepCollection([
-            $this->createStub(StepInterface::class),
+            self::createStub(StepInterface::class),
         ]), ['group-1', 'group-2']);
     }
 }
